Umurzoq_portfolio/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
Pillow==10.2.0